// CurveDentalManagement.UI/src/app/features/doctor/models/doctor.models.ts
export interface Doctor {
    id: string;
    firstName: string;
    lastName: string;
    email: string;
    contactNumber: string;
    specialization: string;
    department: string;
    schedule: string;
    licenseNumber: string;
    yearsOfExperience: string;
    dentalSchool: string;
    officeAddress: string;
    emergencyContact: string;
}